+++ config/env/test.js
@@ -25,6 +25,7 @@ module.exports = {
   },
   maxUploadSize: 10000, // =10kb in bytes. Set ridiculously small just for tests
   port: 3001,
+  illegalStrings: ['trustroots', 'trust', 'roots'],
   app: {
     title: 'Trustroots test environment.',
     description: 'Trustroots test environment.'