+++ modules/tags/client/services/tribes.client.service.js
@@ -1,0 +1,15 @@
+(function() {
+  'use strict';
+
+  angular
+    .module('tags')
+    .factory('TribesService', TribesService);
+
+  /* @ngInject */
+  function TribesService($resource) {
+    return $resource('/api/tribes', {}, {
+      'query': { method: 'GET', isArray: true }
+    });
+  }
+
+})();