+++ CHANGELOG.md
@@ -15,6 +15,16 @@ Versions at `package.json` describe API version.
 
 See [API docs](http://developers.trustroots.org/docs/api/)
 
+#### 0.3.2
+- Add endpoints for tags and tribes:
+  - GET `/api/tribes`
+  - GET `/api/tags`
+  - GET `/api/tags/:tagSlug`
+  - GET `/api/tribes/:tribeSlug`
+  - POST/PUT/DELETE `/api/users/tags`
+  - GET `/api/users/:username` now returns array tags and tribes by `member` object and array of id's by `memberIds` key.
+  - GET `/api/users/:username` doesn't return `public` key anymore for other than authenticated user's profile.
+
 #### 0.3.1
 - Add endpoints for message thread references:
   - GET `/api/references/threads/:userToId`